Working version

+++ src/main/java/com/example/excelchart/domain/CampaignData.java
@@ -5,7 +5,7 @@ import java.math.BigDecimal;
 public class CampaignData {
     private String campaignId;
     private String campaignName;
-    private Double impressions;
+    private Long impressions;
     private Long playouts;
     private Double impPerPlayout;
     private BigDecimal mediaCpm;
@@ -17,7 +17,7 @@ public class CampaignData {
     private BigDecimal clientMargin;
     private BigDecimal totalSpend;
 
-    public CampaignData(String campaignId, String campaignName, Double impressions, Long playouts,
+    public CampaignData(String campaignId, String campaignName, Long impressions, Long playouts,
                         Double impPerPlayout, BigDecimal mediaCpm, BigDecimal totalCpm,
                         BigDecimal mediaCosts, BigDecimal dataCosts, BigDecimal platformCosts,
                         BigDecimal invoiceAmount, BigDecimal clientMargin, BigDecimal totalSpend) {
@@ -39,7 +39,7 @@ public class CampaignData {
     // Getters
     public String getCampaignId() { return campaignId; }
     public String getCampaignName() { return campaignName; }
-    public Double getImpressions() { return impressions; }
+    public Long getImpressions() { return impressions; }
     public Long getPlayouts() { return playouts; }
     public Double getImpPerPlayout() { return impPerPlayout; }
     public BigDecimal getMediaCpm() { return mediaCpm; }
