+++ src/main/java/com/example/excelchart/domain/CityData.java
@@ -6,7 +6,7 @@ import java.math.BigDecimal;
 public class CityData {
     private String campaignId;
     private String city;
-    private Double impressions;
+    private Long impressions;
     private Long playouts;
     private Double impPerPlayout;
     private BigDecimal mediaCpm;
@@ -18,7 +18,7 @@ public class CityData {
     private BigDecimal clientMargin;
     private BigDecimal totalSpend;
 
-    public CityData(String campaignId, String city, Double impressions, Long playouts,
+    public CityData(String campaignId, String city, Long impressions, Long playouts,
                     Double impPerPlayout, BigDecimal mediaCpm, BigDecimal totalCpm,
                     BigDecimal mediaCosts, BigDecimal dataCosts, BigDecimal platformCosts,
                     BigDecimal invoiceAmount, BigDecimal clientMargin, BigDecimal totalSpend) {
@@ -40,7 +40,7 @@ public class CityData {
     // Getters
     public String getCampaignId() { return campaignId; }
     public String getCity() { return city; }
-    public Double getImpressions() { return impressions; }
+    public Long getImpressions() { return impressions; }
     public Long getPlayouts() { return playouts; }
     public Double getImpPerPlayout() { return impPerPlayout; }
     public BigDecimal getMediaCpm() { return mediaCpm; }
